Experimental AtomsBase interface
